+++ src/index.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import './index.css';
 import Twin from './pages/twin';
-import { AuthProvider } from './shared/authContext';
+import { AuthProvider } from './context/authContext';
 
 import { initializeIcons } from '@fluentui/react/lib/Icons';
 initializeIcons();
